debian adaptation

+++ Pi-hole/instal.sh
@@ -26,7 +26,11 @@ server {
 EOF
 
 # Enabling reverse proxy configuration
-sudo ln -s /etc/nginx/sites-available/reverse-proxy /etc/nginx/sites-enabled/reverse-proxy
+if [ ! -L /etc/nginx/sites-enabled/reverse-proxy ]; then
+    sudo ln -s /etc/nginx/sites-available/reverse-proxy /etc/nginx/sites-enabled/reverse-proxy
+else
+    echo "Symbolic link for reverse-proxy already exists."
+fi
 
 # Restarting Nginx server
 echo "restart nginx"
@@ -43,7 +47,7 @@ sudo chmod a+r /etc/apt/keyrings/docker.asc
 
 # Add the repository to Apt sources:
 echo \
-  "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/ubuntu \
+  "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.asc] https://download.docker.com/linux/debian \
   $(. /etc/os-release && echo "$VERSION_CODENAME") stable" | \
   sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
 sudo apt-get update
